update validation errors to take in a FileMap
in theory these errors should only need to know about a single file
at a time during validation. however, we'd like to reuse some of them
(the attribute related ones) during flattening, in which case
into_snippet needs to be able to handle spans in different files
(thus requiring the whole FileMap instead of a single SourceFile).
we'll want to revisit this when we have a better idea of what error
code can be shared.