finalize kind_check's input type
the input type is defined to be a type that is compatible with both
recursive calls into flat::Type (which are always Spanned<Box<Type>>)
and return values from scope.get_type() (which are &Spanned<Type>), at
the cost of some copies of Spans (which should be OK) and some code
to implement From for those two types.
added a TODO to look into if there's a cleaner way to do this, as well
as an explanation in a comment.