cache calcualted kinds
these need to be cached because this allows us to store and dedupe
errors. otherwise, we'd need to recalculate kinds each time we need
to know them during validation, and if it returned an error, we
wouldn't know whether it has been seen or not. by implementing with
a cache, we can do all kindchecking up front and by updating the
cache for types along an errored cycle, we only return each error
once.